securié hors live

+++ src/main/java/fr/studi/promoweb/security/repository/TokenRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.DoubleStream;
 
 @Repository
 public interface TokenRepository extends JpaRepository<Token, Long> {
@@ -16,7 +17,7 @@ public interface TokenRepository extends JpaRepository<Token, Long> {
         on t.user.id = u.id\s
         where u.id = :id and (t.expired = false or t.revoked = false)\s
         """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
 }
